refresh stale search state after mutations
