+++ src/redux/SearchBar.redux/SearchBar.reducer.js
@@ -1,6 +1,4 @@
 import { SEARCH_FILTER, RECIEVE_FILTER } from './SearchBar.types';
-import { searchUsers } from '../../mock-data/mockApi';
-import { searchFilter, recieveFilter } from '../actions';
 
 const initialState = {
   value: '',
@@ -22,25 +20,5 @@ export const searchBarReducer = (state = initialState, action) => {
 
     default:
       return state;
-  }
-};
-
-const LIMIT = 50;
-export const searchFilterFunction = (value) => (dispatch) => {
-  dispatch(searchFilter(value));
-  if (value.length) {
-    const parts = value.split(':').map((e) => e.trim());
-    if (parts.length > 1) {
-      searchUsers(parts[1], {
-        userTypes: [parts[0]],
-        limit: LIMIT,
-      }).then((data) => dispatch(recieveFilter(data)));
-    } else {
-      searchUsers(parts[0], { limit: LIMIT }).then((data) =>
-        dispatch(recieveFilter(data))
-      );
-    }
-  } else {
-    dispatch(recieveFilter({}));
   }
 };
