+++ src/redux/StudentEdit.redux/StudentEdit.actions.js
@@ -1,13 +1,14 @@
 import { editUser as editApi } from '../../mock-data/mockApi';
 import { STUDENT_EDIT_TYPES } from './StudentEdit.types';
+import { refreshSearch } from '../actions';
 
 export const editUser = (id, userData) => (dispatch) => {
-  console.log('userData', userData);
   dispatch({ type: STUDENT_EDIT_TYPES.EDIT_USER_START });
   editApi(id, userData)
-    .then((res) =>
-      dispatch({ type: STUDENT_EDIT_TYPES.EDIT_USER_SUCCESS, payload: res })
-    )
+    .then((res) => {
+      dispatch({ type: STUDENT_EDIT_TYPES.EDIT_USER_SUCCESS, payload: res });
+      dispatch(refreshSearch());
+    })
     .catch((err) =>
       dispatch({ type: STUDENT_EDIT_TYPES.EDIT_USER_FAILURE, payload: err })
     );
